+++ R/cite_bib.R
@@ -11,8 +11,8 @@
 #'   Package to read the bibtex file from.
 #' @param key (`character(1)`)\cr
 #'   Entry of the bibtex file.
-#'   If the key is `"pkg::citation"`, the [citation()] information of the
-#'   package is used instead.
+#'   If the key is `"pkg::citation"`, the [citation()] information of the package is used instead.
+#'   If the package provides multiple citation entries, a specific one can be selected by appending `"::n"` to the string `key` where `n` is the number of the citation entry (defaults to the first entry).
 #'
 #' @return (`character(1)`) Bibentry formated as Rd.
 #'
@@ -31,8 +31,10 @@ cite_bib = function(package, key) {
   assert_string(package)
   assert_string(key)
 
-  if (key == "pkg::citation") {
-    tools::toRd(utils::citation(package))
+  if (grepl("^pkg::citation(::[0-9]+)?", key)) {
+    parts = strsplit(key, "::", fixed = TRUE)[[1L]]
+    nr = if (length(parts) == 2L) 1L else as.integer(parts[3L])
+    tools::toRd(utils::citation(package)[[nr]])
   } else {
     if (!requireNamespace("bibtex", quietly = TRUE)) {
       warningf("Could not load package 'bibtex' to parse citation '%s'", key)